+++ trunk/noiz2sa/src/foe.h
@@ -31,7 +31,7 @@ struct foe {
   Vector pos, vel, ppos, spos, mv;
   int d, spd;
   FoeCommand *cmd;
-  double rank;
+  float rank;
   int spc;
   int type;
   int shield;
@@ -49,12 +49,12 @@ extern "C" {
 
 extern int foeCnt, enNum[];
 
-Foe* addFoe(int x, int y, double rank, int d, int spd, int typek, int shield, 
+Foe* addFoe(int x, int y, float rank, int d, int spd, int typek, int shield, 
 	    BulletMLParser *parser);
-Foe* addFoeBossActiveBullet(int x, int y, double rank, 
+Foe* addFoeBossActiveBullet(int x, int y, float rank, 
 			    int d, int spd, BulletMLParser *state);
-void addFoeActiveBullet(Vector *pos, double rank, 
+void addFoeActiveBullet(Vector *pos, float rank, 
 			int d, int spd, int color, BulletMLState *state);
-void addFoeNormalBullet(Vector *pos, double rank, int d, int spd, int color);
+void addFoeNormalBullet(Vector *pos, float rank, int d, int spd, int color);
 void removeFoe(Foe *fe);
 #endif
